+++ src/lib/internal/helpers/focus.ts
@@ -10,7 +10,6 @@ type HandleFocusArgs = {
 
 export function handleFocus(args: HandleFocusArgs): void {
 	const { prop, defaultEl } = args;
-
 	sleep(1).then(() => {
 		if (prop === undefined) {
 			defaultEl?.focus();
@@ -18,16 +17,15 @@ export function handleFocus(args: HandleFocusArgs): void {
 		}
 
 		const returned = isFunction(prop) ? prop(defaultEl) : prop;
+
 		if (typeof returned === 'string') {
 			// Get el by selector, focus it
-			sleep(1).then(() => {
-				const el = document.querySelector(returned);
-				if (!isHTMLElement(el)) return;
-				el.focus();
-			});
+			const el = document.querySelector(returned);
+			if (!isHTMLElement(el)) return;
+			el.focus();
 		} else if (isHTMLElement(returned)) {
 			// Focus it
-			sleep(1).then(() => returned.focus());
+			returned.focus();
 		}
 	});
 }