+++ src/tests/slider/Slider.spec.ts
@@ -566,3 +566,37 @@ describe.skip('Slider changing options after building', () => {
 		expect(getAllByTestId('tick')).toHaveLength(6);
 	});
 });
+
+describe('Slider (min=0, max=100, step=30)', () => {
+	test('Does not overshoot to the max', async () => {
+		const { getByTestId } = render(Slider, {
+			min: 0,
+			max: 100,
+			step: 30,
+			value: [0],
+		});
+
+		const user = userEvent.setup();
+		const thumb = getByTestId('thumb');
+		await expect(thumb).toBeInTheDocument();
+
+		await act(() => thumb.focus());
+		await user.keyboard(`{${kbd.ARROW_RIGHT}}`);
+		await user.keyboard(`{${kbd.ARROW_RIGHT}}`);
+		await user.keyboard(`{${kbd.ARROW_RIGHT}}`);
+		await user.keyboard(`{${kbd.ARROW_RIGHT}}`);
+
+		expectPercentage({ percentage: 90, thumb, range: getByTestId('range') });
+	});
+
+	test('Autocorrects value outside of steps', () => {
+		const { getByTestId } = render(Slider, {
+			min: 0,
+			max: 100,
+			step: 30,
+			value: [40],
+		});
+
+		expectPercentage({ percentage: 30, thumb: getByTestId('thumb'), range: getByTestId('range') });
+	});
+});